all CRUDS are finished. The remaining thing will be connecting it to the front

+++ src/main/java/co/edu/javeriana/lms/dtos/PaginationMetadataDto.java
@@ -9,7 +9,7 @@ public class PaginationMetadataDto {
     private Integer page;
     private Integer size;
     private Long total;
-    private Long totalPages;
+    private Integer totalPages;
     private String next;
     private String previous;
 }
